rename pid optimization to noise analysis
